New object for generating correct outputs of multiple inputs
